Move subprocess to util
