+++ libopenage/crossplatform/CMakeLists.txt
@@ -1,4 +1,3 @@
 add_sources(libopenage
-	subprocess.cpp
 	os.cpp
 )
